requirements.txt
astroid==2.0.4
Click==7.0
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.1.1
pyobjc==5.1.1
pyobjc-core==5.1.1
pyobjc-framework-Accounts==5.1.1
pyobjc-framework-AddressBook==5.1.1
pyobjc-framework-AppleScriptKit==5.1.1
pyobjc-framework-AppleScriptObjC==5.1.1
pyobjc-framework-ApplicationServices==5.1.1
pyobjc-framework-Automator==5.1.1
pyobjc-framework-AVFoundation==5.1.1
pyobjc-framework-AVKit==5.1.1
pyobjc-framework-CalendarStore==5.1.1
pyobjc-framework-CFNetwork==5.1.1
pyobjc-framework-CloudKit==5.1.1
pyobjc-framework-Cocoa==5.1.1
pyobjc-framework-Collaboration==5.1.1
pyobjc-framework-ColorSync==5.1.1
pyobjc-framework-Contacts==5.1.1
pyobjc-framework-ContactsUI==5.1.1
pyobjc-framework-CoreAudio==5.1.1
pyobjc-framework-CoreAudioKit==5.1.1
pyobjc-framework-CoreBluetooth==5.1.1
pyobjc-framework-CoreData==5.1.1
pyobjc-framework-CoreLocation==5.1.1
pyobjc-framework-CoreMedia==5.1.1
pyobjc-framework-CoreMediaIO==5.1.1
pyobjc-framework-CoreML==5.1.1
pyobjc-framework-CoreServices==5.1.1
pyobjc-framework-CoreSpotlight==5.1.1
pyobjc-framework-CoreText==5.1.1
pyobjc-framework-CoreWLAN==5.1.1
pyobjc-framework-CryptoTokenKit==5.1.1
pyobjc-framework-DictionaryServices==5.1.1
pyobjc-framework-DiscRecording==5.1.1
pyobjc-framework-DiscRecordingUI==5.1.1
pyobjc-framework-DiskArbitration==5.1.1
pyobjc-framework-DVDPlayback==5.1.1
pyobjc-framework-EventKit==5.1.1
pyobjc-framework-ExceptionHandling==5.1.1
pyobjc-framework-ExternalAccessory==5.1.1
pyobjc-framework-FinderSync==5.1.1
pyobjc-framework-FSEvents==5.1.1
pyobjc-framework-GameCenter==5.1.1
pyobjc-framework-GameController==5.1.1
pyobjc-framework-GameKit==5.1.1
pyobjc-framework-GameplayKit==5.1.1
pyobjc-framework-ImageCaptureCore==5.1.1
pyobjc-framework-IMServicePlugIn==5.1.1
pyobjc-framework-InputMethodKit==5.1.1
pyobjc-framework-InstallerPlugins==5.1.1
pyobjc-framework-InstantMessage==5.1.1
pyobjc-framework-Intents==5.1.1
pyobjc-framework-IOSurface==5.1.1
pyobjc-framework-iTunesLibrary==5.1.1
pyobjc-framework-LatentSemanticMapping==5.1.1
pyobjc-framework-LaunchServices==5.1.1
pyobjc-framework-libdispatch==5.1.1
pyobjc-framework-LocalAuthentication==5.1.1
pyobjc-framework-MapKit==5.1.1
pyobjc-framework-MediaAccessibility==5.1.1
pyobjc-framework-MediaLibrary==5.1.1
pyobjc-framework-MediaPlayer==5.1.1
pyobjc-framework-MediaToolbox==5.1.1
pyobjc-framework-ModelIO==5.1.1
pyobjc-framework-MultipeerConnectivity==5.1.1
pyobjc-framework-NetFS==5.1.1
pyobjc-framework-NetworkExtension==5.1.1
pyobjc-framework-NotificationCenter==5.1.1
pyobjc-framework-OpenDirectory==5.1.1
pyobjc-framework-OSAKit==5.1.1
pyobjc-framework-Photos==5.1.1
pyobjc-framework-PhotosUI==5.1.1
pyobjc-framework-PreferencePanes==5.1.1
pyobjc-framework-PubSub==5.1.1
pyobjc-framework-QTKit==5.1.1
pyobjc-framework-Quartz==5.1.1
pyobjc-framework-SafariServices==5.1.1
pyobjc-framework-SceneKit==5.1.1
pyobjc-framework-ScreenSaver==5.1.1
pyobjc-framework-ScriptingBridge==5.1.1
pyobjc-framework-SearchKit==5.1.1
pyobjc-framework-Security==5.1.1
pyobjc-framework-SecurityFoundation==5.1.1
pyobjc-framework-SecurityInterface==5.1.1
pyobjc-framework-ServiceManagement==5.1.1
pyobjc-framework-Social==5.1.1
pyobjc-framework-SpriteKit==5.1.1
pyobjc-framework-StoreKit==5.1.1
pyobjc-framework-SyncServices==5.1.1
pyobjc-framework-SystemConfiguration==5.1.1
pyobjc-framework-VideoToolbox==5.1.1
pyobjc-framework-Vision==5.1.1
pyobjc-framework-WebKit==5.1.1
pyserial==3.4
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.24.0
